+++ karma.conf.js
@@ -90,6 +90,8 @@ module.exports = function(config) {
             // webm-wasm
             {pattern: 'node_modules/webm-wasm/dist/webm-worker.js', included: false, served: true},
             {pattern: 'node_modules/webm-wasm/dist/webm-wasm.wasm', included: false, served: true, type: 'wasm'},
+            // ffmpeg.js
+            {pattern: 'node_modules/ffmpeg.js/ffmpeg-worker-mp4.js', included: false, served: true},
             // gif-recorder: only available on CDN
             'http://cdn.webrtc-experiment.com/gif-recorder.js',