+++ src/js/defaults.js
@@ -113,9 +113,13 @@ const pluginDefaultOptions = {
     animationQuality: 10,
     // Accepts numbers in milliseconds; use this to force intervals-based blobs.
     timeSlice: 0,
-    // Media converter library to use. Legal values are 'ts-ebml' or an empty
-    // string '' to disable (default).
+    // Media converter library to use. Legal values are 'ts-ebml' and 'ffmpeg.js'.
+    // Use an empty string '' to disable (default).
     convertEngine: '',
+    // URL for the converter worker.
+    convertWorkerURL: '',
+    // List of string options to pass to the convert engine.
+    convertOptions: [],
     // Enable keyboard hotkeys.
     hotKeys: false,
     // Use this object to specify additional settings for the library used by the