+++ CHANGES.md
@@ -4,10 +4,13 @@ videojs-record changelog
 3.8.0 - unreleased
 ------------------
 
-- Support for specifying optional plugin settings using the `pluginLibraryOptions`
+- New ffmpeg.js plugin: convert recorded data into other audio/video file formats
+  in the browser (#201)
+- Support for specifying third-party plugin settings using the `pluginLibraryOptions`
   option (#383)
 - New options: `videoBitRate` and `videoFrameRate` (currently only used in the
   webm-wasm plugin)
+- New event: `startConvert` (used in ts-ebml and ffmpeg.js plugins) (#201)
 - Fix `RecordRTC.MediaStreamRecorder` import when using the `timeSlice` option
 - Examples: add `timeSlice` example demonstrating use of `timestamp` event
 - Bump required version for videojs-wavesurfer to 2.9.0 or newer for wavesurfer.js